requirements.txt
martor
gunicorn
requests
python-dotenv
djangorestframework
